requirements.txt
Pillow==5.3.0
django==2.1.2
django-ckeditor==5.6.1